+++ doc/General/ChangeLog.md
@@ -9,7 +9,7 @@ Mirror uses semantic versioning, and the versions shown here are those that were
 - Added: [SyncLists](../Guides/Sync/SyncLists.md) now have Find and FindAll functions.
 - Added: NetworkBehaviour now has OnStopServer and OnStopClient virtual methods
 - Added: Weaver now supports custom Reader & Writer for types in other assemblies
-- Added: Network Manager now has an optional setting to check for and disconnect remote connections that have gone silent
+- Added: Network Manager now has an optional setting to check for and disconnect remote connections that have gone silent for a specified interval.
 - Fixed: NetworkAnimator no longer double-fires SetTrigger / ResetTrigger on the host client
 - Fixed: Destroy is no longer invoked twice on the server for the player object.
 - Changed: NetworkBehaviour: `OnNetworkDestroy` was renamed to `OnStopClient`.